Tweaks to try to get a more consiste user.common.yaml
User.common.yaml is full of traps for the inexerienced user
and this tries to at least catch a few of them, including
an erroneous "=" for the bonding parameter, and implied
eth3 == eth1 based on the parameters in the current file.

This could probably use another cleanup so that IP addresses
are consistent as well (why are some 10. and some 192.)?