Remove empty cisco_coi_user.common.yaml file
The data/hiera_data/vendor/cisco_coi_user.common.yaml file is empty
except for a comment and we presently don't have any real plans
to put anything there.  It does not exist in the StackForge
version of puppet_openstack_builder, and since it's empty
it causes catalog runs to emit a warning of the form:

Warning: Hiera file
/etc/puppet/data/hiera_data/vendor/cisco_coi_user.common.yaml should
contain a hash

This patch fixes the warning by removing the unused file.

Partial-Bug: #1281896